feat: add useLangStore and useDropdown
