+++ src/hooks/use-dropdown.tsx
@@ -1,0 +1,23 @@
+import { useEffect, useRef, useState } from "react";
+
+const useDropdown = () => {
+  const [isOpen, setIsOpen] = useState<boolean>(false);
+  const dropdownRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (dropdownRef.current && !dropdownRef.current.contains(event.target as Node)) {
+        setIsOpen(false);
+      }
+    };
+
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, []);
+
+  return {isOpen, setIsOpen, dropdownRef};
+}
+
+export default useDropdown;
